requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[standard]
psycopg2-binary[standard]
python-dotenv[standard]
sqlalchemy.orm[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart